+++ libft.h
@@ -25,7 +25,7 @@ char	*ft_strdup(char *str);
 char	*ft_strjoin(char const *s1, char const *s2);
 char	*ft_strmap(char const *s, char (*f)(char));
 char	*ft_strmapi(char const *s, char (*f)(unsigned int, char));
-char	*ft_strncat(char *dest, char *src, int nb);
+char	*ft_strncat(char *dest, char *src, size_t nb);
 char	*ft_strncpy(char *dest, char *src, size_t n);
 char	*ft_strnew(size_t size);
 char	*ft_strnstr(char *big, char *little, size_t len);
